refactor(http): rate-limit per key across all surfaces; drop key scope
Rate limiting:
- move the per-request limiter into internal/http (RateLimit) so both the
  client and management APIs share it; mount it on the management routes too.
- key the bucket on the auth method id without a surface prefix (key:<id>), so
  a key has one budget across the client and management APIs.
- rename RATE_LIMIT_CLIENT_PER_MINUTE -> RATE_LIMIT_PER_MINUTE.

Scope:
- public API keys are gone, so WithKey no longer enforces a key scope. The
  per-surface check collapses to enforceSurface: a (private) key is refused on
  the browser-facing client API when the request carries an Origin header.
