+++ internal/http/ratelimit_test.go
@@ -1,4 +1,4 @@
-package oapi
+package http
 
 import (
 	"net/http"
